Tag forwarded spans with custom attributes
Adds a `teleport.forwarded.for` attribute to all spans that are
forwarded to the auth server. This allows consumers of the spans
to identify where the spans are coming from and take possible
action. In some scenarios it may be desirable to drop forwarded
spans along the collection process, by tagging them we can
provide a way for those consumers to identify them. It also
allows for potentially identifying a malicious user that may
be trying to spam the telemetry backend with spans.

Part of #12241